requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
pytz==2018.9
requests==2.21.0
urllib3==1.24.2
